ui: Change Font interface to work with sized glyphs
Glyph metrics depend on whether hinting is performed, so it’s generally impossible to get a size-independent metric and scale it in user code.
The only viable solution seems to not tell the user anything about the glyph until the size is specified; this makes it slightly harder to do layout without rasterizing glyphs.
Additionally this change enables FreeType hinting.